functions/summarize_requirements.txt
functions-framework==3.*
flask==3.*
google-generativeai==0.8.*
google-cloud-bigquery==3.*
